+++ tests/test_pipeline.py
@@ -15,11 +15,9 @@ from pkg_resources import parse_version
 
 from datapipe.types import data_to_index
 from datapipe.compute import Catalog, Pipeline, Table
-from datapipe.core_steps import (
-    do_batch_generate,
-    BatchTransform,
-    DatatableTransformStep,
-)
+from datapipe.step.batch_transform import BatchTransform
+from datapipe.step.datatable_transform import DatatableTransformStep
+from datapipe.step.batch_generate import do_batch_generate
 from datapipe.datatable import DataStore
 from datapipe.store.database import TableStoreDB
 from datapipe.compute import build_compute, run_steps
@@ -57,7 +55,10 @@ def gen_data_df():
     yield pd.DataFrame(
         {
             "id": [f"task_{i}" for i in range(TASKS_COUNT)],
-            "text": [np.random.choice([x for x in string.ascii_letters]) for i in range(TASKS_COUNT)],
+            "text": [
+                np.random.choice([x for x in string.ascii_letters])
+                for i in range(TASKS_COUNT)
+            ],
         }
     )
 
@@ -68,7 +69,9 @@ def wrapped_partial(func, *args, **kwargs):
     return partial_func
 
 
-def convert_to_ls_input_data(data_df, include_preannotations: bool, include_prepredictions: bool):
+def convert_to_ls_input_data(
+    data_df, include_preannotations: bool, include_prepredictions: bool
+):
     columns = ["id", "text"]
 
     for column, bool_ in [
@@ -81,7 +84,9 @@ def convert_to_ls_input_data(data_df, include_preannotations: bool, include_prep
                     {
                         "result": [
                             {
-                                "value": {"choices": [np.random.choice(["Class1", "Class2"])]},
+                                "value": {
+                                    "choices": [np.random.choice(["Class1", "Class2"])]
+                                },
                                 "from_name": "label",
                                 "to_name": "text",
                                 "type": "choices",
@@ -117,7 +122,9 @@ def add_predictions(data_df):
 
 
 INCLUDE_PARAMS = [
-    pytest.param({"include_preannotations": False, "include_prepredictions": False}, id=""),
+    pytest.param(
+        {"include_preannotations": False, "include_prepredictions": False}, id=""
+    ),
     # pytest.param(
     #     {
     #         'include_preannotations': True,
@@ -162,7 +169,10 @@ DELETE_UNANNOTATED_TASKS_ONLY_ON_UPDATE = [
 class CasesLabelStudio:
     @parametrize("include_predictions", INCLUDE_PREDICTIONS)
     @parametrize("include_params", INCLUDE_PARAMS)
-    @parametrize("delete_unannotated_tasks_only_on_update", DELETE_UNANNOTATED_TASKS_ONLY_ON_UPDATE)
+    @parametrize(
+        "delete_unannotated_tasks_only_on_update",
+        DELETE_UNANNOTATED_TASKS_ONLY_ON_UPDATE,
+    )
     def case_ls(
         self,
         include_params,
@@ -311,7 +321,9 @@ def test_ls_moderation(
         df_annotation = ds.get_table("ls_output").get_data()
         for idx in df_annotation.index:
             assert len(df_annotation.loc[idx, "annotations"]) == 1
-            assert df_annotation.loc[idx, "annotations"][0]["result"][0]["value"]["choices"][0] in ["Class1", "Class2"]
+            assert df_annotation.loc[idx, "annotations"][0]["result"][0]["value"][
+                "choices"
+            ][0] in ["Class1", "Class2"]
 
     # Person annotation imitation & incremental processing
     project = get_project_by_title(label_studio_session, project_title)
@@ -346,16 +358,24 @@ def test_ls_moderation(
             label_studio_session.make_request(
                 "POST",
                 f"api/tasks/{task['id']}/annotations/",
-                json=dict(result=annotation["result"], was_cancelled=False, task_id=task["id"]),
+                json=dict(
+                    result=annotation["result"], was_cancelled=False, task_id=task["id"]
+                ),
             )
         run_steps(ds, steps)
-        idxs_df = pd.DataFrame.from_records({"id": [task["data"]["id"] for task in tasks[idxs]]})
-        df_annotation = ds.get_table("ls_output").get_data(idx=data_to_index(idxs_df, ["id"]))
+        idxs_df = pd.DataFrame.from_records(
+            {"id": [task["data"]["id"] for task in tasks[idxs]]}
+        )
+        df_annotation = ds.get_table("ls_output").get_data(
+            idx=data_to_index(idxs_df, ["id"])
+        )
         for idx in df_annotation.index:
-            assert len(df_annotation.loc[idx, "annotations"]) == (1 + include_preannotations)
-            assert df_annotation.loc[idx, "annotations"][0]["result"][0]["value"]["choices"][0] in (
-                ["Class1", "Class2"]
+            assert len(df_annotation.loc[idx, "annotations"]) == (
+                1 + include_preannotations
             )
+            assert df_annotation.loc[idx, "annotations"][0]["result"][0]["value"][
+                "choices"
+            ][0] in (["Class1", "Class2"])
             # if include_predictions:
             #     assert len(df_annotation.loc[idx, 'predictions']) == include_prepredictions + include_predictions
             #     if include_prepredictions or include_predictions:
@@ -383,7 +403,10 @@ def test_ls_when_data_is_changed(
     df1 = pd.DataFrame(
         {
             "id": [f"task_{i}" for i in range(TASKS_COUNT)],
-            "text": (["a", "b", "c", "d", "e", "f", "g", "h", "i", "j"] + ["a"] * (TASKS_COUNT % 10))
+            "text": (
+                ["a", "b", "c", "d", "e", "f", "g", "h", "i", "j"]
+                + ["a"] * (TASKS_COUNT % 10)
+            )
             * (TASKS_COUNT // 10),
         }
     )
@@ -391,7 +414,10 @@ def test_ls_when_data_is_changed(
     df2 = pd.DataFrame(
         {
             "id": [f"task_{i}" for i in range(TASKS_COUNT)],
-            "text": (["A", "B", "C", "d", "E", "f", "G", "h", "I", "j"] + ["a"] * (TASKS_COUNT % 10))
+            "text": (
+                ["A", "B", "C", "d", "E", "f", "G", "h", "I", "j"]
+                + ["a"] * (TASKS_COUNT % 10)
+            )
             * (TASKS_COUNT // 10),
         }
     )
@@ -454,7 +480,10 @@ def test_ls_when_task_is_missing_from_ls(
     df1 = pd.DataFrame(
         {
             "id": [f"task_{i}" for i in range(TASKS_COUNT)],
-            "text": (["a", "b", "c", "d", "e", "f", "g", "h", "i", "j"] + ["a"] * (TASKS_COUNT % 10))
+            "text": (
+                ["a", "b", "c", "d", "e", "f", "g", "h", "i", "j"]
+                + ["a"] * (TASKS_COUNT % 10)
+            )
             * (TASKS_COUNT // 10),
         }
     )
@@ -462,7 +491,10 @@ def test_ls_when_task_is_missing_from_ls(
     df2 = pd.DataFrame(
         {
             "id": [f"task_{i}" for i in range(TASKS_COUNT)],
-            "text": (["A", "B", "C", "d", "E", "f", "G", "h", "I", "j"] + ["a"] * (TASKS_COUNT % 10))
+            "text": (
+                ["A", "B", "C", "d", "E", "f", "G", "h", "I", "j"]
+                + ["a"] * (TASKS_COUNT % 10)
+            )
             * (TASKS_COUNT // 10),
         }
     )
@@ -541,7 +573,11 @@ def test_ls_when_some_data_is_deleted(
         return
     # These steps should upload tasks
     data_df = next(gen_data_df())
-    data_df2 = data_df.set_index("id").drop(index=[f"task_{i}" for i in [0, 3, 5, 7, 9]]).reset_index()
+    data_df2 = (
+        data_df.set_index("id")
+        .drop(index=[f"task_{i}" for i in [0, 3, 5, 7, 9]])
+        .reset_index()
+    )
 
     def _gen():
         yield data_df
@@ -603,7 +639,10 @@ def test_ls_specific_updating_scenary(
     delete_unannotated_tasks_only_on_update: bool,
 ):
     df1 = pd.DataFrame(
-        {"id": [f"task_{i}" for i in range(10)], "text": ["a", "b", "c", "d", "e", "0", "1", "2", "3", "4"]}
+        {
+            "id": [f"task_{i}" for i in range(10)],
+            "text": ["a", "b", "c", "d", "e", "0", "1", "2", "3", "4"],
+        }
     )
 
     df2 = pd.DataFrame(
@@ -630,7 +669,9 @@ def test_ls_specific_updating_scenary(
     project = get_project_by_title(label_studio_session, project_title)
     assert project is not None
     tasks_before = project.get_tasks()
-    tasks_before_sorted = np.array(sorted(tasks_before, key=lambda task: task["data"]["id"]))
+    tasks_before_sorted = np.array(
+        sorted(tasks_before, key=lambda task: task["data"]["id"])
+    )
     assert len(tasks_before) == 10
     tasks_ids_before = [task["id"] for task in tasks_before]
 
@@ -653,7 +694,9 @@ def test_ls_specific_updating_scenary(
         label_studio_session.make_request(
             "POST",
             f"api/tasks/{task['id']}/annotations/",
-            json=dict(result=annotation["result"], was_cancelled=False, task_id=task["id"]),
+            json=dict(
+                result=annotation["result"], was_cancelled=False, task_id=task["id"]
+            ),
         )
 
     # Получаем текущую полученную разметку
@@ -681,7 +724,10 @@ def test_ls_specific_updating_scenary(
     for idx in df_ls.index:
         if df_ls.loc[idx, "id"] in [f"task_{i}" for i in [0, 1, 2, 6, 7, 8]]:
             # Разметка при обновлении задачи не должна уйти, если delete_unannotated_tasks_only_on_update=False
-            if df_ls.loc[idx, "id"] in [f"task_{i}" for i in [0, 1, 2]] and delete_unannotated_tasks_only_on_update:
+            if (
+                df_ls.loc[idx, "id"] in [f"task_{i}" for i in [0, 1, 2]]
+                and delete_unannotated_tasks_only_on_update
+            ):
                 assert len(df_ls.loc[idx, "annotations"]) > 0
                 assert df_ls.loc[idx, "task_id"] in tasks_ids_before
             else: